Praise the Python lord.
1. Download the spreadsheet from google in csv form with name 'google-form.csv'.
This should include all facts from the beginning of time.
2. Run update.py

This will update everything between <!--MARKER_BEGIN--> and <!--MARKER_END--> in
html. The very first fact is excluded from this update because of reasons.

The download.sh.gpg is a convenient script for Lek only because I don't want to
expose the URL to the public. It's just not a true fact.

Praise the python lord again, again, and again.